TG-77: Merged changes from master and deleted extra folders
