--- src/Components/Admin/SuperAdmins/Add/add.module.css
@@ -1,33 +1,0 @@
-.container {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin: 2%;
-}
-
-.formBody {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 50px;
-}
-
-.formRow {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 2%;
-}
-
-.label {
-  font-family: 'Roboto', sans-serif;
-}
-
-.buttons {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
