--- src/Components/Admin/SuperAdmins/Edit/edit.module.css
@@ -1,63 +1,0 @@
-.container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 2%;
-}
-
-.formBody {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 50px;
-}
-
-.formRow {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 2%;
-}
-
-.label {
-  font-family: 'Roboto', sans-serif;
-}
-
-.buttons {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.submit {
-  width: 40%;
-  height: 40%;
-  margin: 5%;
-  padding: 2px 4px 0px 4px;
-  background: #76a068;
-  border-radius: 5px;
-  color: #000000;
-  text-align: center;
-  font-size: 16px;
-  font-family: 'Roboto', sans-serif;
-}
-
-.cancel {
-  width: 40%;
-  height: 40%;
-  margin: 5%;
-  padding: 2px 4px 0px 4px;
-  background: #76a068;
-  border-style: solid;
-  border-width: 2px 2px 2px 2px;
-  border-radius: 5px;
-  border-color: #000000;
-  text-decoration: none;
-  text-align: center;
-  font-size: 16px;
-  color: #000000;
-  font-family: 'Roboto', sans-serif;
-}
