--- src/Components/Admin/SuperAdmins/List/list.module.css
@@ -1,20 +1,0 @@
-.container {
-    margin: 2% 1%;
-}
-
-.table {
-    width: 1000px;
-}
-
-.table, .th {
-    padding-top: 1%;
-    padding-bottom: 1%;
-    border: 1px solid;
-    border-collapse: collapse;
-    font-family: 'Roboto', sans-serif;
-}
-
-.th {
-    background-color: #909090;
-    border: 1px solid;
-}
