--- src/Components/Admin/SuperAdmins/ListItem/listItem.module.css
@@ -1,4 +1,0 @@
-.td {
-    background-color: #ffffff;
-    border: 1px solid;
-}
