--- src/Components/Admins/Edit/edit.module.css
@@ -1,67 +1,0 @@
-.container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 2%;
-}
-
-.formBody {
-  display: flex;
-  flex-direction: column;
-}
-
-.formRow {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 2%;
-}
-
-.label {
-  font-family: 'Roboto', sans-serif;
-}
-
-.buttons {
-  display: flex;
-  flex-direction: row-reverse;
-  justify-content: space-between;
-}
-
-.buttonEditCancel {
-  width: 40%;
-  height: 40%;
-  margin: 5%;
-  padding: 2px 4px 0px 4px;
-  background: #76a068;
-  border-style: solid;
-  border-width: 2px 2px 2px 2px;
-  border-radius: 5px;
-  border-color: #000000;
-  text-decoration: none;
-  text-align: center;
-  font-size: 16px;
-  color: #000000;
-  font-family: 'Roboto', sans-serif;
-}
-
-.back {
-  position: relative;
-  left: 30px;
-  margin: 5%;
-  padding: 2px 4px 0px 4px;
-  border-width: 2px 2px 2px 2px;
-  border-radius: 5px;
-  text-decoration: none;
-  font-size: 16px;
-  font-family: 'Roboto', sans-serif;
-  width: 40%;
-  height: 40%;
-  background: #76a068;
-  border-style: solid;
-  border-color: #000000;
-  color: #000000;
-}
