--- src/Components/Admins/admins.module.css
@@ -1,26 +1,0 @@
-.container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 5%;
-}
-
-.title {
-  font-size: 30px;
-  font-family: 'Roboto', sans-serif;
-  text-transform: uppercase;
-}
-
-.button {
-  width: 80px;
-  height: 80px;
-  background-color: #76a068;
-  border-radius: 50%;
-  border-color: #76a068;
-  box-shadow: 2px 2px 2px grey;
-  text-align: center;
-  font-size: 60px;
-  color: #fff;
-}
