--- src/Components/Employees/employees.module.css
@@ -1,44 +1,0 @@
-.container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 2%;
-}
-
-.container > section {
-  position: fixed;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 5;
-}
-
-.employees {
-  font-size: 30px;
-  font-family: 'Roboto', sans-serif;
-  text-transform: uppercase;
-  text-shadow: 2px 2px 2px grey;
-}
-
-.addbtn {
-  width: 80px;
-  height: 80px;
-  background-color: #76a068;
-  border-radius: 50%;
-  border-color: #76a068;
-  box-shadow: 2px 2px 2px grey;
-  text-transform: uppercase;
-  text-align: center;
-  text-shadow: 2px 2px 2px gray;
-  font-family: 'Roboto', sans-serif;
-  font-size: 50px;
-  color: #ffffff;
-}
