--- src/Components/Employees/index.js
@@ -1,143 +1,0 @@
-import { useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { getEmployees, deleteEmployee } from '../../redux/employees/thunks';
-import styles from './employees.module.css';
-import Preloader from '../Shared/Preloader/Preloader';
-import Table from '../Shared/Table/Table';
-import ModalForm from '../Shared/ModalForm';
-import Form from './Form';
-import ButtonAdd from '../Shared/Buttons/ButtonAdd';
-import ConfirmModal from '../Shared/confirmationModal/confirmModal';
-import SuccessModal from '../Shared/ErrorSuccessModal';
-
-const Employees = () => {
-  const dispatch = useDispatch();
-  const employees = useSelector((state) => state.employees.list);
-  const isLoading = useSelector((state) => state.employees.isLoading);
-  const [showModalFormAdd, setShowModalFormAdd] = useState(false);
-  const [showModalFormEdit, setShowModalFormEdit] = useState(false);
-  const [showModalFormDelete, setShowModalFormDelete] = useState(false);
-  const [showSuccessModal, setShowSuccessModal] = useState(false);
-  const [response, setResponse] = useState('');
-  const [employeeId, setEmployeeId] = useState();
-
-  useEffect(() => {
-    dispatch(getEmployees());
-  }, []);
-
-  const closeModalFormAdd = () => {
-    setShowModalFormAdd(false);
-  };
-
-  const closeModalFormEdit = () => {
-    setShowModalFormEdit(false);
-  };
-
-  let modalEdit;
-  if (showModalFormEdit) {
-    modalEdit = (
-      <ModalForm isOpen={showModalFormEdit} handleClose={closeModalFormEdit} title="Edit Employee">
-        <Form
-          closeModalForm={closeModalFormEdit}
-          edit={true}
-          item={employees.find((item) => item._id === employeeId)}
-        />
-      </ModalForm>
-    );
-  }
-
-  let modalAdd;
-  if (showModalFormAdd) {
-    modalAdd = (
-      <ModalForm isOpen={showModalFormAdd} handleClose={closeModalFormAdd} title="Add Employee">
-        <Form closeModalForm={closeModalFormAdd} />
-      </ModalForm>
-    );
-  }
-
-  let modalDelete;
-  if (showModalFormDelete) {
-    modalDelete = (
-      <ConfirmModal
-        isOpen={showModalFormDelete}
-        handleClose={() => {
-          setShowModalFormDelete(false);
-        }}
-        confirmDelete={() => {
-          dispatch(deleteEmployee(employeeId, setResponse)).then(() => {
-            setShowModalFormDelete(false);
-            setShowSuccessModal(true);
-          });
-        }}
-        title="Delete Employee"
-        message={'Are you sure you want to delete this employee?'}
-      />
-    );
-  }
-
-  return isLoading &&
-    !showModalFormEdit &&
-    !showModalFormAdd &&
-    !showModalFormDelete &&
-    !showSuccessModal ? (
-    <Preloader />
-  ) : (
-    <section className={styles.container}>
-      <h2 className={styles.employees}> Employees </h2>
-      <Table
-        data={employees}
-        headers={[
-          '_id',
-          'firstName',
-          'lastName',
-          'phone',
-          'email',
-          'active',
-          'isProjectManager',
-          'projects',
-          'timeSheets'
-        ]}
-        titles={[
-          'ID',
-          'Name',
-          'Surname',
-          'Phone',
-          'Email',
-          'Active',
-          'Project Manager',
-          'Projects',
-          'TimeSheets'
-        ]}
-        delAction={(id) => {
-          setEmployeeId(id);
-          setShowModalFormDelete(true);
-        }}
-        editAction={(id) => {
-          setEmployeeId(id);
-          setShowModalFormEdit(true);
-        }}
-      />
-      {modalEdit}
-      {modalAdd}
-      {modalDelete}
-      {isLoading ? <Preloader /> : null}
-      <ButtonAdd
-        clickAction={() => {
-          setShowModalFormAdd(true);
-        }}
-      />
-      <SuccessModal
-        show={showSuccessModal}
-        closeModal={() => {
-          setShowSuccessModal(false);
-        }}
-        closeModalForm={() => {
-          setShowSuccessModal(false);
-        }}
-        successResponse={response}
-      />
-    </section>
-  );
-};
-
-export default Employees;
