--- src/Components/Projects/Form/AddMember/addMember.module.css
@@ -1,77 +1,0 @@
-.container {
-  margin: 20px auto;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 300px;
-}
-
-.container label {
-  color: #373867;
-  display: block;
-  padding-bottom: 3px;
-  text-align: justify;
-  font-weight: 700;
-  font-size: 14px;
-}
-
-.inputs {
-  font-family: 'Roboto', sans-serif;
-  padding: 4px 6px;
-  width: 100%;
-  height: 25px;
-  border: 0.5px solid #373867;
-  border-radius: 3px;
-  color: #373867;
-}
-
-.container button {
-  margin-top: 15px;
-  padding: 7px 20px;
-  background: #76a068;
-  border-radius: 3.7px;
-  border: none;
-  font-weight: 700;
-  font-size: 14px;
-  color: #373867;
-  cursor: pointer;
-}
-.divcontainer {
-  margin: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.divcontainer h2 {
-  text-align: center;
-  width: 700px;
-  margin-bottom: 20px;
-  padding: 15px 0;
-  background-color: #373867;
-  border-radius: 15px 15px 0 0;
-  font-weight: 700;
-  font-size: 30px;
-  line-height: 35px;
-  color: #ffffff;
-  text-transform: uppercase;
-  text-align: center;
-}
-.loading {
-  margin: auto;
-}
-
-.divcontainer a {
-  display: block;
-  margin: 15px auto;
-  padding: 10px;
-  background: #373867;
-  border-radius: 3.7px;
-  border: none;
-  font-weight: 700;
-  text-decoration: none;
-  font-size: 16px;
-  color: #ffffff;
-  cursor: pointer;
-}
