--- src/Components/Projects/Form/ListMembers/listmembers.module.css
@@ -1,37 +1,0 @@
-.container {
-  margin: 10px;
-  display: flex;
-  flex-direction: column;
-  height: 200px;
-  overflow-y: scroll;
-  border: solid 0.5px;
-  border-radius: 3px;
-  font-size: 14px;
-  align-items: center;
-}
-
-.container h3 {
-  color: #373867;
-}
-
-.container table {
-  margin: 0 10px;
-  border-spacing: 5px;
-}
-
-.container th {
-  margin: 0 5px;
-  text-align: left;
-}
-.container a,
-.addmember {
-  display: block;
-  margin: auto;
-  padding: 7px;
-  background: #76a068;
-  border-radius: 5px;
-  font-weight: 500;
-  font-size: 12px;
-  text-decoration: none;
-  color: #373867;
-}
