--- src/Components/Projects/projects.module.css
@@ -1,30 +1,0 @@
-.container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 2%;
-}
-
-.projects {
-  font-size: 30px;
-  font-family: 'Roboto', sans-serif;
-  text-transform: uppercase;
-  text-shadow: 2px 2px 2px grey;
-}
-
-.addbtn {
-  width: 80px;
-  height: 80px;
-  background-color: #76a068;
-  border-radius: 50%;
-  border-color: #76a068;
-  box-shadow: 2px 2px 2px grey;
-  text-transform: uppercase;
-  text-align: center;
-  text-shadow: 2px 2px 2px gray;
-  font-family: 'Roboto', sans-serif;
-  font-size: 50px;
-  color: #ffffff;
-}
