--- src/Components/Tasks/Form/form.module.css
@@ -1,11 +1,0 @@
-.form {
-  flex: 1;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  align-items: flex-end;
-  align-content: space-between;
-  row-gap: 10px;
-  overflow: auto;
-}
