--- src/Components/TimeSheets/EditItem/EditItem.module.css
@@ -1,66 +1,0 @@
-.container {
-  position: fixed;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  width: 100%;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  overflow: auto;
-  background-color: rgba(0, 0, 0, 0.5);
-}
-
-.container label {
-  display: flex;
-  flex-direction: column;
-}
-
-.container form {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 35%;
-  height: 70%;
-  background-color: rgb(255, 255, 255);
-  border-radius: 10px;
-  gap: 15px;
-}
-
-.container input[type='submit'] {
-  padding: 10px;
-  background-color: #76a068;
-  border: black solid 1px;
-  border-radius: 5px;
-  font-size: 15px;
-  color: #373867;
-}
-
-.container input[type='submit']:hover {
-  background-color: rgb(55, 120, 28);
-  color: rgb(255, 255, 255);
-  transition-duration: 0, 4s;
-}
-
-.container input[type='submit']:active {
-  background-color: rgb(38, 82, 19);
-  transform: translateY(4px);
-}
-
-.closeButton {
-  padding: 10px 18px;
-  background-color: #76a068;
-  border: black solid 1px;
-  border-radius: 5px;
-  font-size: 15px;
-  color: #373867;
-}
-
-.closeButton:hover {
-  background-color: rgb(55, 120, 28);
-  color: rgb(255, 255, 255);
-  transition-duration: 0, 4s;
-}
