--- src/Components/TimeSheets/Modal/Modal.module.css
@@ -1,54 +1,0 @@
-.container {
-  position: fixed;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  width: 100%;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  overflow: auto;
-  background-color: rgba(0, 0, 0, 0.5);
-}
-
-.modal {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: flex-start;
-  width: 30%;
-  min-height: 15%;
-  background-color: rgb(255, 255, 255);
-  border-radius: 5px;
-  font-weight: 900;
-  overflow: hidden;
-}
-
-.tittle {
-  background: #232f3e;
-  width: 100%;
-  color: #fff;
-  padding: 15px;
-  font-size: 20px;
-  font-family: 'Roboto', sans-serif;
-  font-weight: 400;
-}
-
-.description {
-  display: flex;
-  align-self: flex-start;
-  color: #373867;
-  font-weight: 600;
-  padding: 40px 15px;
-}
-
-.btnClose {
-  background-color: #76a068;
-  font-size: 16px;
-  border-radius: 5px;
-  padding: 10px 20px;
-  align-self: center;
-  margin-bottom: 15px;
-}
