--- src/Components/TimeSheets/time-sheets.module.css
@@ -1,54 +1,0 @@
-.container {
-  flex: 1;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  flex-direction: column;
-}
-
-.containerLoading {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.loader {
-  border: 10px solid #f3f3f3;
-  border-top: 10px solid #76a068;
-  border-radius: 50%;
-  width: 80px;
-  height: 80px;
-  animation: spin 0.7s linear infinite;
-}
-
-.addButton {
-  margin-bottom: 20px;
-  width: 80px;
-  height: 80px;
-  background-color: #76a068;
-  border-radius: 50%;
-  border-color: #76a068;
-  box-shadow: 2px 2px 2px grey;
-  text-transform: uppercase;
-  text-align: center;
-  text-shadow: 2px 2px 2px grey;
-  font-family: 'Roboto', sans-serif;
-  font-size: 50px;
-  color: #ffffff;
-}
-
-.title {
-  margin-top: 50px;
-  margin-bottom: 50px;
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
