+++ src/Routes/employee.js
@@ -3,7 +3,7 @@ import { Switch, Route, useRouteMatch } from 'react-router-dom';
 import Home from 'Components/Employee/Home';
 import Timesheet from 'Components/Employee/Timesheet';
 import Projects from 'Components/Employee/Projects';
-// import Profile from 'Components/Employee/EditProfile';
+import Profile from 'Components/Employee/EditProfile';
 
 const Employee = () => {
   const { url } = useRouteMatch();
@@ -11,7 +11,7 @@ const Employee = () => {
     <Switch>
       <Route path={`${url}/time-sheet`} component={Timesheet} />
       <Route path={`${url}/projects`} component={Projects} />
-      {/* <Route path={`${url}/profile`} component={Profile} /> */}
+      <Route path={`${url}/profile`} component={Profile} />
       <Route exact path={`${url}/`} component={Home} />
     </Switch>
   );
